requirements.txt
sanic==22.6.2
min-dalle==0.4.7
IPython
pillow